+++ addons/website/__manifest__.py
@@ -118,8 +118,10 @@
     'assets': {
         'web.assets_frontend': [
             ('replace', 'web/static/src/legacy/js/public/public_root_instance.js', 'website/static/src/js/content/website_root_instance.js'),
+            'website/static/src/libs/zoomodoo/zoomodoo.scss',
             'website/static/src/scss/website.scss',
             'website/static/src/scss/website.ui.scss',
+            'website/static/src/libs/zoomodoo/zoomodoo.js',
             'website/static/src/js/utils.js',
             'website/static/src/js/tours/tour_utils.js',
             'website/static/src/js/content/website_root.js',
