+++ debian/copyright
@@ -378,7 +378,7 @@ Files: addons/web/static/src/js/libs/download.js
 Copyright: 2016 dandavis
 License: MIT
 
-Files: addons/web/static/src/js/libs/zoomodoo.js
+Files: addons/website/static/src/libs/zoomodoo/zoomodoo.js
 Copyright: 2013 Matt Hinchliffe
 License: MIT
